Reorganized compare_models into subdirectory and cleaned up config code
